code refactor: css changes refactored
