+++ src/features/taskPlanner/taskModal/TaskModal.module.css
@@ -16,6 +16,7 @@
 
 .actionButton {
   text-transform: none !important;
+  background-color: #713fff !important;
 }
 
 .titleWrapper {
@@ -103,6 +104,7 @@
   padding-bottom: 5px;
   border-radius: 8px;
   cursor: pointer;
+  margin-right: 10px;
 }
 
 .low {
@@ -112,7 +114,7 @@
   padding-right: 20px;
   padding-top: 5px;
   padding-bottom: 5px;
-  border-radius: 8px; 
+  border-radius: 8px;
   cursor: pointer;
 }
 
@@ -149,6 +151,8 @@
   position: flex-end;
   display: flex;
   font-size: 12;
+  display: flex;
+  justify-content: right;
 }
 
 .addButton:disabled {
@@ -162,4 +166,9 @@
   background-color: #713fff;
   text-transform: none;
   float: inline-end;
+}
+
+.actionButtons {
+  display: flex;
+  justify-content: right;
 }
